MB-2964 Move old apps code to tasks package
