--- apps/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-from .anon import AnonBehavior  # noqa
-from .service_member import ServiceMemberUserBehavior  # noqa
-from .office import OfficeUserBehavior  # noqa
